refactor: move burrego runtime to dedicated module
This allows a better organization of the code
