+++ src/runtimes/mod.rs
@@ -1,9 +1,9 @@
-pub mod burrego;
+pub(crate) mod rego;
 pub(crate) mod wapc;
 pub(crate) mod wasi_cli;
 
 pub(crate) enum Runtime {
     Wapc(wapc::WapcStack),
-    Burrego(burrego::BurregoStack),
+    Burrego(rego::BurregoStack),
     Cli(wasi_cli::Stack),
 }
